Added support for .env-files

+++ index.js
@@ -1,12 +1,16 @@
+require('dotenv').config();
 const parseString = require('xml2js').parseString,
   fs = require('fs'),
   path = require('path'),
-  filePath = path.join(__dirname, 'rekordbox.xml'),
-  outputFolderName = 'Playlists',
+  xmlFileName = process.env.XML_FILENAME ?? 'rekordbox.xml',
+  xmlFilePath = path.join(__dirname, xmlFileName),
+  outputFolderName = process.env.OUTPUT_FOLDER_NAME ?? 'Playlists',
   outputFolderPath = path.join(__dirname, outputFolderName),
-  dryRun = false;
+  useSubfolders = process.env.USE_SUBFOLDERS ? process.env.USE_SUBFOLDERS === 'true' : true, // Default true
+  dryRun = process.env.DRY_RUN === 'true',
+  lineDelimiter = process.env.LINE_DELIMITER ?? "\n";
 
-console.log('Reading file: ' + filePath);
+console.log('Reading file: ' + xmlFilePath);
 
 function resolvePlaylistName(element, parentElement) {
   const nameParts = [];
@@ -30,51 +34,66 @@ function handleLevel(tracks, element, parentElement) {
     return;
   }
 
-  const playlistName = resolvePlaylistName(element, parentElement);
+  const playlistName = resolvePlaylistName(element, useSubfolders ? null : parentElement);
   const tracksInPlaylist = element.TRACK;
 
-  const tracksInFile = [];
-  const fileName = playlistName + '.m3u';
-  const filePath = path.join(__dirname, outputFolderName, fileName);
-
   if (tracksInPlaylist) {
+    const tracksInFile = [];
+    const fileName = `${playlistName}.m3u`;
+    let filePath;
+
+    if (useSubfolders) {
+      const subFolderName = parentElement ? parentElement.$.Name : '';
+      const subFolderPath = path.join(__dirname, outputFolderName, subFolderName);
+      filePath = path.join(subFolderPath, fileName);
+
+      if (!fs.existsSync(subFolderPath)) {
+        fs.mkdirSync(subFolderPath, { recursive: true });
+      }
+    } else {
+      filePath = path.join(__dirname, outputFolderName, fileName);
+    }
+    
     Object.values(tracksInPlaylist).forEach(track => {
       const trackId = track.$.Key;
       const trackObject = tracks.find(t => t.$.TrackID === trackId);
       const trackPath = decodeURIComponent(trackObject.$.Location.replace('file://localhost', ''));
       tracksInFile.push(trackPath);
     });
-  }
 
-  if (dryRun) {
-    console.log(`The playlist "${playlistName}" would have been saved to file "${fileName}" with "${tracksInPlaylist.length}" tracks.`);
-  } else {
-    fs.writeFile(filePath, tracksInFile.join("\n"), function(err) {
-      if(err) {
-          return console.log(err);
-      }
-      console.log(`The playlist "${playlistName}" was saved to file "${fileName}" with "${tracksInPlaylist.length}" tracks.`);
-    });
+    const resultMessage = `${tracksInPlaylist.length} tracks in playlist "${playlistName}" ${dryRun ? 'would have been' : 'was'} written to file "${fileName}"`;
+    if (dryRun) {
+      console.log(resultMessage);
+    } else {
+      fs.writeFile(filePath, tracksInFile.join(lineDelimiter), function(err) {
+        if(err) {
+            return console.log(err);
+        }
+        console.log(resultMessage);
+      });
+    }
   }
-  
 }
 
+if (!fs.existsSync(xmlFilePath)) {
+  console.log('XML-file not found. Please check filename and path.');
+  return;
+}
 
 if (fs.existsSync(outputFolderPath)) {
   fs.rmSync(outputFolderPath, { recursive: true, force: true });
 }
 fs.mkdirSync(outputFolderPath);
 
-fs.readFile(filePath, {encoding: 'utf-8'}, function(err, data){
-    if (!err) {
-      
+fs.readFile(xmlFilePath, {encoding: 'utf-8'}, function(error, data) {
+    if (!error) {
       parseString(data, function (err, result) {
         const tracks = result.DJ_PLAYLISTS.COLLECTION[0].TRACK;
         const playLists = result.DJ_PLAYLISTS.PLAYLISTS[0].NODE[0].NODE
         playLists.forEach(playList => handleLevel(tracks, playList));
       });
     } else {
-        console.log(err);
+        console.log('An error occurred:', error);
     }
 });
